% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_score.R
\name{normalize_score}
\alias{normalize_score}
\title{Normalize numeric values to a given range}
\usage{
normalize_score(x, old_min = 1, old_max = 5, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{Numeric vector of values to scale}

\item{old_min}{Minimum score (e.g 1 for a rating between 1 and 5)}

\item{old_max}{Maximum score (e.g 5 for a rating between 1 and 5)}

\item{new_min}{Desired minimum (default = 0)}

\item{new_max}{Desired maximum (default = 1)}
}
\value{
Vector scaled to [new_min, new_max]
}
\description{
Normalize numeric values to a given range
}
\examples{
ratings <- c(3.5, 4.2, 4.8)
normalize_score(ratings, old_min = 1, old_max = 5)

}
